Rename user info service as server-only
