+++ frontend/app/routes/_gcweb-app._index.tsx
@@ -3,7 +3,7 @@ import { Link, useLoaderData } from '@remix-run/react';
 
 import { useTranslation } from 'react-i18next';
 
-import { userService } from '~/services/user-service';
+import { userService } from '~/services/user-service.server';
 import { getEnv } from '~/utils/env.server';
 import { getTypedI18nNamespaces } from '~/utils/locale-utils';
